chore: Moved `HighNeedsHistoryParametersValidator` into correct namespace
